// src/app/models/course-management/courseAttachData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CourseAttachData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** PK auto run (course_attach)  */
    courseAttachId?: number;
    /** FK (course_main) */
    courseId?: number;
    /** FileName */
    filename?: string;
    /** ชื่อไฟล์ (ภาษาอังกฤษ) */
    fileNameEn?: string;
    /** ชื่อไฟล์ (ภาษาไทย) */
    fileNameTh?: string;
    first?: number;
    mode?: string;
    /** Module */
    module?: number;
    /** Prefix */
    prefix?: string;
    rowNum?: number;
    size?: number;
}